% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hcat.execute}
\alias{hcat.execute}
\title{hcat.execute}
\usage{
hcat.execute(hqlQuery, database="default", outputDir="", fieldDelimiter="\t")
}
\value{
Results as
		a data frame if no output directory is specified; or
		writes results to a file in HDFS in the specified output directory with the specified fieldDelimiter
}
\description{
Executes a HQL Query against an object in HCatalog
}

