% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hcat.gettabledirectory}
\alias{hcat.gettabledirectory}
\title{hcat.gettabledirectory}
\usage{
hcat.gettabledirectory(tablename, database="default")
}
\value{
HFDS Directory for a HCatalog Table
}
\description{
Returns the HDFS directory for a HCatalog Table
}

